m airtime changes

+++ controllers/services.mairtime.helpers.controller.js
@@ -28,12 +28,12 @@ const mtnVTUTopUp = async (user,trxRef,time,service,phoneNumber,amount,res)=>{
   let serviceCharge = serviceCategory.serviceCharge;
   let predifinedDiscount = service.discount;
   let discount = await helpers.getDiscount(user.id,predifinedDiscount);
-  let totalAmount = parseFloat(amount) + parseFloat(serviceCharge); 
+  let totalAmount = parseInt(amount) + parseInt(serviceCharge); 
   if(discount){
     totalAmount = totalAmount  - discount;
   }
   let profit = totalAmount - amount;
-  let walletBalance = parseFloat(wallet.accountBalance);
+  let walletBalance = parseInt(wallet.accountBalance);
   if(walletBalance < totalAmount){
     const transaction = await models.transaction.create(
       {
@@ -109,12 +109,12 @@ const airtimePurchase = async (user,trxRef,time,service,phoneNumber,amount,res)=
   let serviceCharge = serviceCategory.serviceCharge;
   let predifinedDiscount = service.discount;
   let discount = await helpers.getDiscount(user.id,predifinedDiscount);
-  let totalAmount = parseFloat(amount) + parseFloat(serviceCharge); 
+  let totalAmount = parseInt(amount) + parseInt(serviceCharge); 
   if(discount){
     totalAmount = totalAmount  - discount;
   }
   let profit = totalAmount - amount;
-  let walletBalance = parseFloat(wallet.accountBalance);
+  let walletBalance = parseInt(wallet.accountBalance);
   if(walletBalance < totalAmount){
     const transaction = await models.transaction.create(
       {
@@ -190,12 +190,12 @@ const foreignAirtimePurchase = async (user,trxRef,time,service,phoneNumber,amoun
   let serviceCharge = serviceCategory.serviceCharge;
   let predifinedDiscount = service.discount;
   let discount = await helpers.getDiscount(user.id,predifinedDiscount);
-  let totalAmount = parseFloat(amountInNaira) + parseFloat(serviceCharge); 
+  let totalAmount = parseInt(amountInNaira) + parseInt(serviceCharge); 
   if(discount){
     totalAmount = totalAmount  - discount;
   }
   let profit = totalAmount - amount;
-  let walletBalance = parseFloat(wallet.accountBalance);
+  let walletBalance = parseInt(wallet.accountBalance);
   if(walletBalance < totalAmount){
     const transaction = await models.transaction.create(
       {
@@ -272,12 +272,12 @@ const mtnDataGifting = async (user,trxRef,time,service,phoneNumber,amount,dataSi
   let serviceCharge = serviceCategory.serviceCharge;
   let predifinedDiscount = service.discount;
   let discount = await helpers.getDiscount(user.id,predifinedDiscount);
-  let totalAmount = parseFloat(amount) + parseFloat(serviceCharge); 
+  let totalAmount = parseInt(amount) + parseInt(serviceCharge); 
   if(discount){
     totalAmount = totalAmount  - discount;
   }
   let profit = totalAmount - amount;
-  let walletBalance = parseFloat(wallet.accountBalance);
+  let walletBalance = parseInt(wallet.accountBalance);
   if(walletBalance < totalAmount){
     const transaction = await models.transaction.create(
       {
@@ -353,12 +353,12 @@ const mtnDataShare = async (user,trxRef,time,service,phoneNumber,amount,dataSize
   let serviceCharge = serviceCategory.serviceCharge;
   let predifinedDiscount = service.discount;
   let discount = await helpers.getDiscount(user.id,predifinedDiscount);
-  let totalAmount = parseFloat(amount) + parseFloat(serviceCharge); 
+  let totalAmount = parseInt(amount) + parseInt(serviceCharge); 
   if(discount){
     totalAmount = totalAmount  - discount;
   }
   let profit = totalAmount - amount;
-  let walletBalance = parseFloat(wallet.accountBalance);
+  let walletBalance = parseInt(wallet.accountBalance);
   if(walletBalance < totalAmount){
     const transaction = await models.transaction.create(
       {
@@ -434,12 +434,12 @@ const dataTopUp = async (user,trxRef,time,service,phoneNumber,amount,res)=>{
   let serviceCharge = serviceCategory.serviceCharge;
   let predifinedDiscount = service.discount;
   let discount = await helpers.getDiscount(user.id,predifinedDiscount);
-  let totalAmount = parseFloat(amount) + parseFloat(serviceCharge); 
+  let totalAmount = parseInt(amount) + parseInt(serviceCharge); 
   if(discount){
     totalAmount = totalAmount  - discount;
   }
   let profit = totalAmount - amount;
-  let walletBalance = parseFloat(wallet.accountBalance);
+  let walletBalance = parseInt(wallet.accountBalance);
   if(walletBalance < totalAmount){
     const transaction = await models.transaction.create(
       {
@@ -515,12 +515,12 @@ const purchaseElectricity = async (user,trxRef,time,service,amount,serviceId,met
   let serviceCharge = serviceCategory.serviceCharge;
   let predifinedDiscount = service.discount;
   let discount = await helpers.getDiscount(user.id,predifinedDiscount);
-  let totalAmount = parseFloat(amount) + parseFloat(serviceCharge); 
+  let totalAmount = parseInt(amount) + parseInt(serviceCharge); 
   if(discount){
     totalAmount = totalAmount  - discount;
   }
   let profit = totalAmount - amount;
-  let walletBalance = parseFloat(wallet.accountBalance);
+  let walletBalance = parseInt(wallet.accountBalance);
   if(walletBalance < totalAmount){
     const transaction = await models.transaction.create(
       {
@@ -596,12 +596,12 @@ const waecPinPurchase = async (user,trxRef,time,service,amount,res)=>{
   let serviceCharge = serviceCategory.serviceCharge;
   let predifinedDiscount = service.discount;
   let discount = await helpers.getDiscount(user.id,predifinedDiscount);
-  let totalAmount = parseFloat(amount) + parseFloat(serviceCharge); 
+  let totalAmount = parseInt(amount) + parseInt(serviceCharge); 
   if(discount){
     totalAmount = totalAmount  - discount;
   }
   let profit = totalAmount - amount;
-  let walletBalance = parseFloat(wallet.accountBalance);
+  let walletBalance = parseInt(wallet.accountBalance);
   if(walletBalance < totalAmount){
     const transaction = await models.transaction.create(
       {
@@ -674,12 +674,12 @@ const necoPinPurchase = async (user,trxRef,time,service,amount,res)=>{
   let serviceCharge = serviceCategory.serviceCharge;
   let predifinedDiscount = service.discount;
   let discount = await helpers.getDiscount(user.id,predifinedDiscount);
-  let totalAmount = parseFloat(amount) + parseFloat(serviceCharge); 
+  let totalAmount = parseInt(amount) + parseInt(serviceCharge); 
   if(discount){
     totalAmount = totalAmount  - discount;
   }
   let profit = totalAmount - amount;
-  let walletBalance = parseFloat(wallet.accountBalance);
+  let walletBalance = parseInt(wallet.accountBalance);
   if(walletBalance < totalAmount){
     const transaction = await models.transaction.create(
       {
@@ -752,12 +752,12 @@ const tvRecharge = async (user,trxRef,time,service,amount,cardNo,customerName,in
   let serviceCharge = serviceCategory.serviceCharge;
   let predifinedDiscount = service.discount;
   let discount = await helpers.getDiscount(user.id,predifinedDiscount);
-  let totalAmount = parseFloat(amount) + parseFloat(serviceCharge); 
+  let totalAmount = parseInt(amount) + parseInt(serviceCharge); 
   if(discount){
     totalAmount = totalAmount  - discount;
   }
   let profit = totalAmount - amount;
-  let walletBalance = parseFloat(wallet.accountBalance);
+  let walletBalance = parseInt(wallet.accountBalance);
   if(walletBalance < totalAmount){
     const transaction = await models.transaction.create(
       {
@@ -836,12 +836,12 @@ const startimesRecharge = async (user,trxRef,time,service,amount,cardNo,phoneNum
   let serviceCharge = serviceCategory.serviceCharge;
   let predifinedDiscount = service.discount;
   let discount = await helpers.getDiscount(user.id,predifinedDiscount);
-  let totalAmount = parseFloat(amount) + parseFloat(serviceCharge); 
+  let totalAmount = parseInt(amount) + parseInt(serviceCharge); 
   if(discount){
     totalAmount = totalAmount  - discount;
   }
   let profit = totalAmount - amount;
-  let walletBalance = parseFloat(wallet.accountBalance);
+  let walletBalance = parseInt(wallet.accountBalance);
   if(walletBalance < totalAmount){
     const transaction = await models.transaction.create(
       {
